+++ src/d/d_bg_w_kcol.cpp
@@ -7,6 +7,7 @@
 
 #include <algorithm>
 #include <cmath>
+#include <stdint.h>
 
 #include "SSystem/SComponent/c_bg_s_shdw_draw.h"
 #include "SSystem/SComponent/c_m2d.h"
@@ -553,8 +554,8 @@ bool dBgWKCol::GroundCross(cBgS_GndChk* i_chk) {
         int sp20 = 4 * (((u32)sp34 >> shift) << m_pkc_head->m_area_xy_blocks_shift |
                         ((u32)sp30 >> shift) << m_pkc_head->m_area_x_blocks_shift |
                          (u32)sp38 >> shift);
-        while ((sp20 = (*(int*)((int)block + sp20))) >= 0) {
-            block = (int)block + sp20;
+        while ((sp20 = (*(int*)(block + sp20))) >= 0) {
+            block = block + sp20;
             shift--;
             sp20 = (((u32)sp34 >> shift & 1) << 2 |
                     ((u32)sp30 >> shift & 1) << 1 |
@@ -695,15 +696,15 @@ void dBgWKCol::ShdwDraw(cBgS_ShdwDraw* param_0) {
 
                         int x_sp44 = minX_spA0;
                         do {
-                            u32 block_sp40 = (u32)m_pkc_head->m_block_data;
+                            uintptr_t block_sp40 = (uintptr_t)m_pkc_head->m_block_data;
                             u32 shift_sp3C = m_pkc_head->m_block_width_shift;
                             int offset_sp38 =
                                 4 * (((u32)z_sp4C >> shift_sp3C) << m_pkc_head->m_area_xy_blocks_shift |
                                      ((u32)y_sp48 >> shift_sp3C) << m_pkc_head->m_area_x_blocks_shift |
                                       (u32)x_sp44 >> shift_sp3C);
 
-                            while ((offset_sp38 = *(u32*)((int)block_sp40 + offset_sp38)) >= 0) {
-                                block_sp40 = (int)block_sp40 + offset_sp38;
+                            while ((offset_sp38 = *(u32*)(block_sp40 + offset_sp38)) >= 0) {
+                                block_sp40 = block_sp40 + offset_sp38;
                                 shift_sp3C--;
                                 offset_sp38 = (((u32)z_sp4C >> shift_sp3C & 1) << 2 |
                                                ((u32)y_sp48 >> shift_sp3C & 1) << 1 |
@@ -2007,7 +2008,7 @@ bool dBgWKCol::SplGrpChk(dBgS_SplGrpChk* param_0) {
 
     bool sp09 = false;
     do {
-        u32 sp28 = (u32)m_pkc_head->m_block_data;
+        uintptr_t sp28 = (uintptr_t)m_pkc_head->m_block_data;
         u32 sp24 = m_pkc_head->m_block_width_shift;
         int sp20 = 4 * (((u32)sp34 >> sp24) << m_pkc_head->m_area_xy_blocks_shift |
                         ((u32)sp2C >> sp24) << m_pkc_head->m_area_x_blocks_shift |
